requirements.txt
absl-py==2.3.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
firebase-admin==5.3.0
Flask==3.1.1
flatbuffers==25.2.10
gast==0.6.0
google-api-core==2.10.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.7.2
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.46.0
gunicorn==23.0.0
h5py==3.13.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.1.0
networkx==3.5
numba==0.61.2
numpy==2.1.3
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
# pyheif==0.8.0
PyMatting==1.1.14
pyparsing==3.2.3
referencing==0.36.2
rembg==2.0.66
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.19.0
termcolor==3.1.0
tifffile==2025.6.1
tqdm==4.67.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2